Reformed and refactored code structure
