+++ novobi_addons/omni_address_checker/models/__init__.py
@@ -1,4 +1,3 @@
-from . import fedex_request
 from . import delivery_fedex
 from . import stock_picking
 from . import res_partner
