+++ novobi_addons/omni_address_checker/utils/__init__.py
@@ -1,0 +1,1 @@
+from . import fedex_request
